Rem previous encryption for SMS login :bug:
When added the SMS password login, some problems ocurred with the
encryption used. I've tried to use new methods for password work how I
wanted, but no progress had made on this.

I've manage to get SMS login working. But what I really wanted was to
use SMS as a confirmation of user, not as login. The login should be
made confortably with the username and password of user.