Mod name of project :snowflake:
I've stumbled upon a problem with Cordova, that's commond besides lots
of Javascript projects. But specially with Cordova, to work on Android,
the compilation process of the APK for testing on smarphones, requires
lots of file operations.

Unfortunately, whether Cordova, Google or NPM fault, some of these move,
remove, copy operations hits Windows limit for paths. And for these
project to run on Windows 10, I've tried a lot of some hacks. The main
one, was involving gpedit.msc, a file to manage privac settings on
Windows, but my version of Windows couldn't modify NSTF properties.

The only solution left to me, was moving the project one level ahead
on my computer, and also change the name of project itself.
*tracking-sanitation-ufcg* it's a long name for a project. I've
simplified on a Ctrl-F substitution to *tsu*. That way, anyone coding
on this project will delay more the problem related to Windows paths.

I need to change *tsu* to *tracking-sanitation-ufcg* or a better name
for the project. For a while, I'm gonna maintain these shorter name.