Fix upload of images :bug:
After deciding to store images using dataurl strins on MongoDB, I've
manage to get the upload working. Unfortunately, as I tried to work on
thumbnail generation, the upload stopped working because of a collection
permissions problems.

The problem was solved, through some changes on publish options, and
subscribing images and thumbs collection on the top of the project, at
the client side.