Fix display of images to scroll content :bug:
I've fixed a problem, involving the interface. The contents of the
records view, wasn't scrollable, and the pagination control were on top
of the content. Also, the cards were overflowing on top of another.

Now the pagination control it's fixed, and the cards can scroll without
one staying on top of another.