Fix offline support on app :bug:
I've managed to work a way, to make the app completely offline usable.

I've used a new package to store user session offline. Also, I've could
store insertions and removes offline for sync. I've changed the way the
images are stored, so it can be possible to store them offline.

Unfortunately, for security purposes, the app still needs an server,
where I'm planning to use Heroku on that.